Bug 1246061.  r=jandem, r=bz, r=luke, r=froydnj
UltraBlame original commit: 967dcb05f34702b5fdbc12892c3f92d67fc3450a

+++ js/src/jsfriendapi.h
@@ -1248,7 +1248,7 @@ struct ExpandoAndGeneration {
   }
 
   JS::Heap<JS::Value> expando;
-  uint32_t generation;
+  uint64_t generation;
 };
 
 typedef enum DOMProxyShadowsResult {
